refactor(widgets): move VisibilityPill into its own file beside its test
VisibilityPill (and its getVisibilityPillConfig helper) lived in Card.tsx while
its spec sat in VisibilityPill.cy.tsx. Match the established pattern of a
component next to its Cypress test (WelcomeBanner, ScoreSummaryChart, ...):
extract it to widgets/VisibilityPill.tsx, import it into Card, and point the
spec at the new file. Pure move — no behavior change; Card drops the now-unused
Badge and react-icons/fi imports.
